app/requirements.txt
Django==4.1.10
djangorestframework==3.14.0
psycopg2-binary==2.9.6
drf-yasg==1.21.7
django-filter==23.2